requirements.txt
djangorestframework==3.11.1
cassandra-driver==3.24.0
django==3.1
twilio==6.45.0
Pillow==7.2.0
pyfcm==1.4.7
channels==2.4.0
channels_redis==2.4.2
django-cassandra-engine==1.5.5
psycopg2==2.8.4
django-debug-toolbar==2.2
django-redis-sessions==0.3